Gemini/2.LLM_&_Large_Image_model/requirements.txt
streamlit
google-generativeai 
python-dotenv
ipykernel
